[feat] 댓글 기능 추가, 서비스 권한 검사 방식 변경

+++ src/main/java/org/example/schedulemanagementapp/domain/schedule/service/ScheduleServiceImpl.java
@@ -1,15 +1,24 @@
 package org.example.schedulemanagementapp.domain.schedule.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.schedulemanagementapp.domain.comment.entity.Comment;
 import org.example.schedulemanagementapp.domain.schedule.dto.ScheduleBaseRequest;
 import org.example.schedulemanagementapp.domain.schedule.dto.ScheduleBaseResponse;
 import org.example.schedulemanagementapp.domain.schedule.dto.SchedulerUpdateRequest;
 import org.example.schedulemanagementapp.domain.schedule.entity.Schedule;
 import org.example.schedulemanagementapp.domain.schedule.repository.ScheduleRepository;
+import org.example.schedulemanagementapp.domain.user.entity.User;
+import org.example.schedulemanagementapp.domain.user.repository.UserRepository;
+import org.example.schedulemanagementapp.domain.user.service.UserService;
+import org.example.schedulemanagementapp.global.exception.CustomException;
+import org.example.schedulemanagementapp.global.exception.ErrorCode;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -21,14 +30,16 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ScheduleServiceImpl implements ScheduleService {
     private final ScheduleRepository scheduleRepository;
+    private final UserService userService;
 
     @Transactional
     @Override
-    public ScheduleBaseResponse save(ScheduleBaseRequest dto) {
+    public ScheduleBaseResponse save(Long userId, ScheduleBaseRequest requestDto) {
+        User user = userService.findUserByIdOrThrow(userId);
         Schedule schedule = new Schedule(
-                dto.getUserId(),
-                dto.getTitle(),
-                dto.getContent()
+                user,
+                requestDto.getTitle(),
+                requestDto.getContent()
         );
 
         Schedule savedSchedule = scheduleRepository.save(schedule);
@@ -52,13 +63,15 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Transactional
     @Override
-    public ScheduleBaseResponse updateScheduleById(Long scheduleId, SchedulerUpdateRequest dto) {
+    public ScheduleBaseResponse updateScheduleById(Long userId, Long scheduleId, SchedulerUpdateRequest requestDto) {
         Schedule schedule = findScheduleByIdOrThrow(scheduleId);
+        // 유저 검증
+        if (!CheckUserIdInSchedule(schedule, userId))
+            throw new CustomException(ErrorCode.NO_PERMISSION);
 
         schedule.update(
-                dto.getUserId() != null ? dto.getUserId() : schedule.getUserId(),
-                dto.getTitle() != null ? dto.getTitle() : schedule.getTitle(),
-                dto.getContent() != null ? dto.getContent() : schedule.getContent()
+                Optional.ofNullable(requestDto.getTitle()).orElse(schedule.getTitle()),
+                Optional.ofNullable(requestDto.getContent()).orElse(schedule.getContent())
         );
 
         return ScheduleBaseResponse.of(schedule);
@@ -66,20 +79,31 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Transactional
     @Override
-    public void deleteScheduleById(Long scheduleId) {
+    public void deleteScheduleById(Long userId, Long scheduleId) {
         Schedule schedule = findScheduleByIdOrThrow(scheduleId);
+        // 유저 검증
+        if (!CheckUserIdInSchedule(schedule, userId))
+            throw new CustomException(ErrorCode.NO_PERMISSION);
+
         scheduleRepository.delete(schedule);
     }
 
-    /**
-     * 레포지토리에서 아이디를 찾아 일정 엔티티를 반환하고 없다면 예외 발생
-     *
-     * @param scheduleId 일정 아이디
-     * @return 일정 엔티티
-     */
-    private Schedule findScheduleByIdOrThrow(Long scheduleId) {
+    @Transactional(readOnly = true)
+    @Override
+    public Schedule findScheduleByIdOrThrow(Long scheduleId) {
         return scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new IllegalArgumentException("Schedule Not Found!")
+                () -> new CustomException(ErrorCode.SCHEDULE_NOT_FOUND)
         );
     }
+
+    /**
+     * 일정 작성자 아이디를 확인하고 비교값 반환
+     *
+     * @param schedule 일정 엔티티
+     * @param userId   유저 아이디
+     */
+    private boolean CheckUserIdInSchedule(Schedule schedule, Long userId) {
+        return Objects.equals(schedule.getUser().getId(), userId);
+    }
+
 }
